Allow terminating luns for a specific pool in SCALE
