+++ src/middlewared/middlewared/plugins/iscsi_/global_base.py
@@ -11,11 +11,5 @@ class GlobalActionsBase(ServicePartBase):
        """
 
     @private
-    async def alua_enabled(self):
-        """
-        Returns whether iSCSI ALUA is enabled or not.
-        """
-
-    @private
     async def terminate_luns_for_pool(self, pool_name):
         raise NotImplementedError()
