+++ src/middlewared/middlewared/plugins/iscsi_/global_freebsd.py
@@ -43,21 +43,6 @@ class ISCSIGlobalService(Service, GlobalActionsBase):
             })
         return filter_list(sessions, filters, options)
 
-    async def alua_enabled(self):
-        """
-        Returns whether iSCSI ALUA is enabled or not.
-        """
-        if await self.middleware.call('system.is_freenas'):
-            return False
-        if not await self.middleware.call('failover.licensed'):
-            return False
-
-        license = (await self.middleware.call('system.info'))['license']
-        if license and 'FIBRECHANNEL' in license['features']:
-            return True
-
-        return (await self.middleware.call('iscsi.global.config'))['alua']
-
     async def terminate_luns_for_pool(self, pool_name):
         cp = await run(['ctladm', 'devlist', '-b', 'block', '-x'], check=False, encoding='utf8')
         for lun in etree.fromstring(cp.stdout).xpath('//lun'):
