+++ src/middlewared/middlewared/plugins/iscsi_/iscsi_global.py
@@ -78,3 +78,18 @@ class ISCSIGlobalService(SystemServiceService):
             await self.middleware.call('etc.generate', 'loader')
 
         return await self.config()
+
+    async def alua_enabled(self):
+        """
+        Returns whether iSCSI ALUA is enabled or not.
+        """
+        if await self.middleware.call('system.is_freenas'):
+            return False
+        if not await self.middleware.call('failover.licensed'):
+            return False
+
+        license = (await self.middleware.call('system.info'))['license']
+        if license and 'FIBRECHANNEL' in license['features']:
+            return True
+
+        return (await self.middleware.call('iscsi.global.config'))['alua']
