+++ src/middlewared/middlewared/plugins/pool.py
@@ -1466,8 +1466,8 @@ class PoolService(CRUDService):
                 pass
             else:
                 raise
-        if osc.IS_FREEBSD:
-            await self.middleware.call('iscsi.global.terminate_luns_for_pool', pool['name'])
+
+        await self.middleware.call('iscsi.global.terminate_luns_for_pool', pool['name'])
 
         job.set_progress(30, 'Removing pool disks from swap')
         disks = [i async for i in await self.middleware.call('pool.get_disks', oid)]
